utils: Rename enums_context_processors to context_processors
Let's add more context processors in there.
